RM snapshot

+++ src/main/java/org/renci/requestmanager/LinkRequestInfo.java
@@ -14,14 +14,14 @@ public class LinkRequestInfo {
     
     private String wfUuid; // workflow unique id
     private String linkId; // link id to be used by entity requesting the link
-    private int linkBandwidth; // bandwidth in Mb/s; say to the stitchport
+    private long linkBandwidth; // bandwidth in Mb/s; say to the stitchport
     
     // following is the identifier in the stitchport registry; 
     // querying with this identifier gives the url of the stitchport mapper service; 
     // querying that service gives IP level info, vlan tag, port etc.
     private String stitchPortID;
 
-    public LinkRequestInfo(String wfUuid, String linkId, int linkBandwidth, String stitchPortID) {
+    public LinkRequestInfo(String wfUuid, String linkId, long linkBandwidth, String stitchPortID) {
         this.wfUuid = wfUuid;
         this.linkId = linkId;
         this.linkBandwidth = linkBandwidth;
@@ -55,11 +55,11 @@ public class LinkRequestInfo {
         this.linkId = linkId;
     }
 
-    public int getLinkBandwidth() {
+    public long getLinkBandwidth() {
         return linkBandwidth;
     }
 
-    public void setLinkBandwidth(int linkBandwidth) {
+    public void setLinkBandwidth(long linkBandwidth) {
         this.linkBandwidth = linkBandwidth;
     }
